requirements.txt
streamlit
psycopg2-binary 
python-dotenv
openai
streamlit_pills
beautifulsoup4
pandas
matplotlib
scipy
nltk
seaborn